r16886: Add the mapping from CIFS posix locks to system POSIX locks. That's the easy part... Next will come adding the unlock ! Jeremy.

+++ source/include/locking.h
@@ -80,7 +80,7 @@ struct lock_struct {
 	struct lock_context context;
 	br_off start;
 	br_off size;
-	int fnum;
+	uint16 fnum;
 	enum brl_type lock_type;
 	enum brl_flavour lock_flav;
 };
